requirements.txt
colorama==0.3.7
coverage==4.2
nose==1.3.7
py==1.4.31
PyHamcrest==1.9.0
pyserial==3.1.1
pytest==2.9.2
pytest-cov==2.3.0
python-termstyle==0.1.10
rednose==1.2.1
six==1.10.0
